Warn users when they try to resolve a local reference without precomputing scopes
